pyproject.toml
[tool.poetry]
name = "accountfinance"
version = "0.1"
description = ""
authors = ["Sergey Popov <sergei.popov174@gmail.com>", "Назаров Всеволод <mr.seva99@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
sqlalchemy = "^1.3.18"
python-dotenv = "^0.13.0"
marshmallow = "^3.6.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
